FILE STRUCT UPDATE 🗃
