+++ src/components/Announcement/index.js
@@ -1,0 +1,4 @@
+import AnnouncementPanel from './AnnouncementPanel'
+import CreateAnnouncement from './CreateAnnouncement'
+
+export { AnnouncementPanel, CreateAnnouncement }
