+++ src/components/Form/index.js
@@ -1,0 +1,4 @@
+import CreateForm from './CreateForm'
+import FillForm from './FillForm'
+
+export { CreateForm, FillForm }
