--- src/components/compIndex.js
@@ -1,13 +1,0 @@
-import AnnouncementPanel from './AnnouncementPanel/AnnouncementPanel'
-import CreateAnnouncement from './CreateAnnouncement/CreateAnnouncement'
-import CreateForm from './CreateForm/CreateForm'
-import FormField from './CreateForm/FormField'
-import FillForm from './FillForm/FillForm'
-
-export {
-  AnnouncementPanel,
-  CreateAnnouncement,
-  CreateForm,
-  FormField,
-  FillForm,
-}
