+++ src/views/Dashboard/StudentDash.jsx
@@ -1,30 +1,28 @@
 import React from 'react'
 import { useEffect } from 'react'
 import { useHistory } from 'react-router-dom'
-import { AnnouncementPanel } from '../../../components/compIndex'
 import AppBar from '@mui/material/AppBar'
 import Box from '@mui/material/Box'
 import Toolbar from '@mui/material/Toolbar'
 import Typography from '@mui/material/Typography'
 import Button from '@mui/material/Button'
 import IconButton from '@mui/material/IconButton'
-
 import {
   getRole,
   removeSession,
-} from '../../../services/LocalStorageService/LocalStorageService'
+} from '../../services/LocalStorageService/LocalStorageService'
+import { AnnouncementPanel } from '../../components/Announcement'
 
-function FacultyDash() {
+function StudentDash() {
   const history = useHistory()
   useEffect(() => {
-    let checkAuth = getRole() === '1'
+    let checkAuth = getRole() === '2'
     if (!checkAuth) {
       removeSession()
       history.push('/login')
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [])
-
   return (
     <div>
       <Box sx={{ flexGrow: 1 }}>
@@ -40,7 +38,7 @@ function FacultyDash() {
               {/* <MenuIcon /> */}
             </IconButton>
             <Typography variant="h6" component="div" sx={{ flexGrow: 1 }}>
-              Faculty
+              Student
             </Typography>
             <Button
               color="inherit"
@@ -54,10 +52,9 @@ function FacultyDash() {
           </Toolbar>
         </AppBar>
       </Box>
-      <h1> Announcement </h1>
       <AnnouncementPanel />
     </div>
   )
 }
 
-export default FacultyDash
+export default StudentDash
