+++ src/views/Login/Login.jsx
@@ -58,13 +58,18 @@ function Login() {
   }
 
   return (
-    <div className="container d-flex justify-content-center align-items-center"
-      style={{display: 'flex',  justifyContent:'center', alignItems:'center', height: '80vh'}}
+    <div
+      className="container d-flex justify-content-center align-items-center"
+      style={{
+        display: 'flex',
+        justifyContent: 'center',
+        alignItems: 'center',
+        height: '80vh',
+      }}
     >
       <form
         className="login-form card mt-5"
         style={{ width: '25em' }}
-        
         onSubmit={handleSubmit(onSubmit, onError)}
       >
         <h2 className="card-header text-center">Log in</h2>
