+++ src/views/index.js
@@ -1,8 +1,8 @@
+import AdminDash from './Dashboard/AdminDash'
 import Dashboard from './Dashboard/Dashboard'
+import FacultyDash from './Dashboard/FacultyDash'
+import StudentDash from './Dashboard/StudentDash'
 import Home from './Home/Home'
 import Login from './Login/Login'
-import AdminDash from './Roles/Admin/AdminDash'
-import FacultyDash from './Roles/Faculty/FacultyDash'
-import StudentDash from './Roles/Student/StudentDash'
 
 export { Home, Login, Dashboard, AdminDash, FacultyDash, StudentDash }
